pyproject.toml
[project]
name = "composer-trade-common"
version = "0.1.0"
description = "A collection of common groups used on Composer symphonies. To be used in combination with composer-trade-py SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "composer-trade-py>=0.1.2",
]
